Fixed timing issues while upgrading the packages
To fix this issue, the packages must ensure that the binaries are
stopped before upgrading the packages. Also, to avoid some issues
related to the upgrade of SCA files, I have changed the temporary
directory of those files, to avoid get deleted by any possible
restart of the binaries while upgrading.